有符号无符号比较  修正
git-svn-id: svn://192.168.2.188/qtcms@538 894f9759-fa2a-45cf-9159-0a04a166245e

+++ src/functionlibs/Recorder/Recorder.cpp
@@ -249,7 +249,7 @@ void Recorder::run()
 				quint64 TotalNumberOfFreeBytes;
 				m_StorageMgr.GetDiskFreeSpaceEx(sdisk.toAscii().data(),&FreeByteAvailable,&TotalNumberOfBytes,&TotalNumberOfFreeBytes);
 
-				if( TotalNumberOfFreeBytes<= m_StorageMgr.getFreeSizeForDisk() * 1024 * 1024)
+				if( TotalNumberOfFreeBytes<= (quint64)m_StorageMgr.getFreeSizeForDisk() * 1024 * 1024)
 				{ // not enough free space
 
 					qDebug("Not enough free space\n");
